[#1] 파일 분리
