+++ Hyodor/Hyodor/Shared/Models/SharedPhoto.swift
@@ -4,6 +4,7 @@
 //
 //  Created by 김상준 on 4/24/25.
 //
+
 import Foundation
 
 // MARK: - 공유 앨범의 사진 모델 (전체 조회/동기화 등에서 사용)
@@ -22,10 +23,5 @@ struct SharedPhoto: Identifiable, Codable {
        }
 }
 
-
-// /api/gallery/all 응답
-struct AllPhotosResponse: Codable {
-    let photos: [SharedPhoto]
-}
